# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones del sistema SistaL.",
        "title": "MS Operations API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
...
        "/operaciones/dashboard": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene resumen de operaciones para dashboard (hoy, pendientes, alertas, actividad reciente)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operaciones"
                ],
                "summary": "Dashboard de operaciones",
                "responses": {
                    "200": {
                        "description": "Datos del dashboard",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/operaciones/estadisticas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene estadísticas de peticiones, despachos y facturas con filtros opcionales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operaciones"
                ],
                "summary": "Obtener estadísticas de operaciones",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "name": "fecha_desde",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "name": "fecha_hasta",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de sucursal",
                        "name": "id_sucursal",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas completas",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de autenticación JWT en formato Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}